+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+## Version 1.0.0-RC2
+_2016-05-20_
+- **[BUG]** Fix a bug: IM Content was not an array as expected.
+- **[IMPROVEMENT]** Add simple example for IM Messages.
+
 ## Version 1.0.0-RC1
 _2016-05-12_
-- Initial release
+- Initial release candidate.